Add remote control support to panasonic_viera integration
